+++ source/Plugins/Imports/uPSI_API_Task.pas
@@ -14,8 +14,8 @@ uses
   ,Classes
   ,uPSComponent
   ,uPSRuntime
-  ,uPSCompiler
-  ;
+  ,uPSCompiler,
+  ProjectTask;
 
 type
 (*----------------------------------------------------------------------------*)