+++ source/TagType.pas
@@ -1,0 +1,14 @@
+unit TagType;
+
+interface
+
+Type
+  TTagType = (ttUnknown, ttWorkingDirectory, ttProperties);
+
+
+implementation
+
+
+
+
+end.