+++ source/TokenProcessor.pas
@@ -1,0 +1,115 @@
+unit TokenProcessor;
+
+interface
+
+Uses SysUtils, Classes, API_Output;
+
+type
+
+   tTokenProcessor = class(TStringList)
+   private
+     fiTokenIndex: Integer;
+     foOutput: tAPI_Output;
+   protected
+   public
+     constructor Create; overload;
+     constructor Create(aOutput: tAPI_Output); overload;
+
+     function ToString: String; override;
+
+     function GetFirstToken: string;
+     function GetNextToken(aIgnoreNextToken: Boolean = false): string; overload;
+     function GetNextToken(aTokenIndex: Integer): string; overload;
+     function IsNextTokenEquals: boolean;
+     function IsNextTokenOpenBracket: boolean;
+
+     function EOF: Boolean;
+
+     property TokenIndex: Integer
+         read fiTokenIndex
+         write fiTokenIndex;
+
+     property oOutput: tAPI_Output
+       read foOutput
+       write foOutput;
+   end;
+
+implementation
+
+// Token Processor
+constructor tTokenProcessor.Create;
+begin
+  fiTokenIndex:= 0;
+end;
+
+constructor tTokenProcessor.Create(aOutput: tAPI_Output);
+begin
+  fiTokenIndex:= 0;
+
+  foOutput := aOutput;
+end;
+
+function tTokenProcessor.EOF: Boolean;
+begin
+  Result := (fiTokenIndex >= Count);
+end;
+
+function tTokenProcessor.GetNextToken(aIgnoreNextToken: Boolean): String;
+begin
+  Result := '';
+
+  if fiTokenIndex > Count then  Exit;
+
+  if Count > 0 then
+    Result := Trim(Strings[fiTokenIndex]);
+
+  if Not aIgnoreNextToken then
+    begin
+      Inc(fiTokenIndex);
+    end;
+end;
+
+
+function tTokenProcessor.GetFirstToken: string;
+begin
+  fiTokenIndex:=0;
+  if Count =0 then Exit;
+  Result := Trim(Strings[fiTokenIndex]);
+  Inc(fiTokenIndex);
+end;
+
+
+function tTokenProcessor.GetNextToken(aTokenIndex: Integer): string;
+begin
+  Result := '';
+
+  fiTokenINdex := aTokenIndex;
+
+  if aTokenIndex > Count then  Exit;
+
+  if Count > 0 then
+    Result := Trim(Strings[aTokenIndex]);
+end;
+
+
+function tTokenProcessor.IsNextTokenEquals: boolean;
+begin
+  Result := GetNextToken = '=';
+end;
+
+function tTokenProcessor.IsNextTokenOpenBracket: Boolean;
+begin
+  Result := GetNextToken = '(';
+end;
+
+function tTokenProcessor.ToString: String;
+Var
+  I: Integer;
+begin
+  Result := '';
+
+  for I := 0 to Self.Count -1 do
+     Result := Result +Trim(Strings[i]);
+end;
+
+end.